--- Tests/MockedTests/MockImplementations/MockAsyncThrowingImplementationTests.swift
@@ -1,51 +1,0 @@
-//
-//  MockAsyncThrowingImplementationTests.swift
-//  MockedTests
-//
-//  Created by Gray Campbell on 12/8/23.
-//
-
-import XCTest
-@testable import Mocked
-
-final class MockAsyncThrowingImplementationTests: XCTestCase {
-
-    // MARK: Typealiases
-
-    typealias SUT = MockAsyncThrowingImplementation<ReturnValue>
-    typealias ReturnValue = Int
-
-    // MARK: Call As Function Tests
-
-    func testReturnsCallAsFunction() async throws {
-        let sut: SUT = .returns { 5 }
-
-        let returnValue = try await sut(description: self.description())
-
-        XCTAssertEqual(returnValue, 5)
-    }
-
-    func testThrowsCallAsFunction() async throws {
-        let sut: SUT = .throws { URLError(.badURL) }
-
-        do {
-            _ = try await sut(description: self.description())
-            XCTFail("Expected sut to throw an error")
-        } catch let error as URLError {
-            XCTAssertEqual(error.code, .badURL)
-        } catch {
-            XCTFail("Expected \(error) to equal URLError(.badURL)")
-        }
-    }
-}
-
-// MARK: - Helpers
-
-extension MockAsyncThrowingImplementationTests {
-    private func description() -> MockImplementationDescription {
-        MockImplementationDescription(
-            type: Self.self,
-            member: "sut"
-        )
-    }
-}
