+++ Tests/MockedTests/MockMethods/MockVoidMethods/MockVoidAsyncThrowingMethodWithParametersTests.swift
@@ -15,182 +15,178 @@ final class MockVoidAsyncThrowingMethodWithParametersTests: XCTestCase {
     typealias SUT = MockVoidAsyncThrowingMethodWithParameters<Arguments>
     typealias Arguments = (string: String, boolean: Bool)
 
+    // MARK: Implementation Tests
+
+    func testImplementationDefaultValue() async {
+        let (sut, _) = SUT.makeMethod()
+
+        guard case .unimplemented = sut.implementation else {
+            XCTFail("Expected implementation to equal .unimplemented")
+            return
+        }
+    }
+
     // MARK: Call Count Tests
 
     func testCallCount() async throws {
-        try await self.withSUT { sut, invoke in
-            XCTAssertEqual(sut.callCount, .zero)
+        let (sut, invoke) = SUT.makeMethod()
 
-            try await invoke(("a", true))
-            XCTAssertEqual(sut.callCount, 1)
+        XCTAssertEqual(sut.callCount, .zero)
 
-            sut.error = URLError(.badURL)
+        try await invoke(("a", true))
+        XCTAssertEqual(sut.callCount, 1)
 
-            do {
-                try await invoke(("b", false))
-                XCTFail("Expected invoke to throw an error")
-            } catch let error as URLError {
-                XCTAssertEqual(error.code, .badURL)
-            } catch {
-                XCTFail("Expected \(error) to equal URLError(.badURL)")
-            }
+        sut.implementation = .throws(URLError(.badURL))
 
-            XCTAssertEqual(sut.callCount, 2)
+        do {
+            try await invoke(("b", false))
+            XCTFail("Expected invoke to throw an error")
+        } catch let error as URLError {
+            XCTAssertEqual(error.code, .badURL)
+        } catch {
+            XCTFail("Expected \(error) to equal URLError(.badURL)")
         }
+
+        XCTAssertEqual(sut.callCount, 2)
     }
 
     // MARK: Invocations Tests
 
     func testInvocations() async throws {
-        try await self.withSUT { sut, invoke in
-            XCTAssertTrue(sut.invocations.isEmpty)
+        let (sut, invoke) = SUT.makeMethod()
 
-            try await invoke(("a", true))
-            XCTAssertEqual(sut.invocations.count, 1)
+        XCTAssertTrue(sut.invocations.isEmpty)
+
+        try await invoke(("a", true))
+        XCTAssertEqual(sut.invocations.count, 1)
+        XCTAssertEqual(sut.invocations.first?.string, "a")
+        XCTAssertEqual(sut.invocations.first?.boolean, true)
+
+        sut.implementation = .throws(URLError(.badURL))
+
+        do {
+            try await invoke(("b", false))
+            XCTFail("Expected invoke to throw an error")
+        } catch let error as URLError {
+            XCTAssertEqual(error.code, .badURL)
+            XCTAssertEqual(sut.invocations.count, 2)
             XCTAssertEqual(sut.invocations.first?.string, "a")
             XCTAssertEqual(sut.invocations.first?.boolean, true)
-
-            sut.error = URLError(.badURL)
-
-            do {
-                try await invoke(("b", false))
-                XCTFail("Expected invoke to throw an error")
-            } catch let error as URLError {
-                XCTAssertEqual(error.code, .badURL)
-                XCTAssertEqual(sut.invocations.count, 2)
-                XCTAssertEqual(sut.invocations.first?.string, "a")
-                XCTAssertEqual(sut.invocations.first?.boolean, true)
-                XCTAssertEqual(sut.invocations.last?.string, "b")
-                XCTAssertEqual(sut.invocations.last?.boolean, false)
-            } catch {
-                XCTFail("Expected \(error) to equal URLError(.badURL)")
-            }
+            XCTAssertEqual(sut.invocations.last?.string, "b")
+            XCTAssertEqual(sut.invocations.last?.boolean, false)
+        } catch {
+            XCTFail("Expected \(error) to equal URLError(.badURL)")
         }
     }
 
     // MARK: Last Invocation Tests
 
     func testLastInvocation() async throws {
-        try await self.withSUT { sut, invoke in
-            XCTAssertNil(sut.lastInvocation)
+        let (sut, invoke) = SUT.makeMethod()
 
-            try await invoke(("a", true))
-            XCTAssertEqual(sut.lastInvocation?.string, "a")
-            XCTAssertEqual(sut.lastInvocation?.boolean, true)
+        XCTAssertNil(sut.lastInvocation)
 
-            sut.error = URLError(.badURL)
+        try await invoke(("a", true))
+        XCTAssertEqual(sut.lastInvocation?.string, "a")
+        XCTAssertEqual(sut.lastInvocation?.boolean, true)
 
-            do {
-                try await invoke(("b", false))
-                XCTFail("Expected invoke to throw an error")
-            } catch let error as URLError {
-                XCTAssertEqual(error.code, .badURL)
-                XCTAssertEqual(sut.lastInvocation?.string, "b")
-                XCTAssertEqual(sut.lastInvocation?.boolean, false)
-            } catch {
-                XCTFail("Expected \(error) to equal URLError(.badURL)")
-            }
+        sut.implementation = .throws(URLError(.badURL))
+
+        do {
+            try await invoke(("b", false))
+            XCTFail("Expected invoke to throw an error")
+        } catch let error as URLError {
+            XCTAssertEqual(error.code, .badURL)
+            XCTAssertEqual(sut.lastInvocation?.string, "b")
+            XCTAssertEqual(sut.lastInvocation?.boolean, false)
+        } catch {
+            XCTFail("Expected \(error) to equal URLError(.badURL)")
         }
     }
 
     // MARK: Thrown Errors Tests
 
     func testThrownErrors() async throws {
-        try await self.withSUT { sut, invoke in
-            XCTAssertTrue(sut.thrownErrors.isEmpty)
+        let (sut, invoke) = SUT.makeMethod()
 
-            try await invoke(("a", true))
-            XCTAssertTrue(sut.thrownErrors.isEmpty)
+        XCTAssertTrue(sut.thrownErrors.isEmpty)
 
-            sut.error = URLError(.badURL)
+        try await invoke(("a", true))
+        XCTAssertTrue(sut.thrownErrors.isEmpty)
 
-            do {
-                try await invoke(("b", false))
-                XCTFail("Expected invoke to throw an error")
-            } catch let error as URLError {
-                XCTAssertEqual(error.code, .badURL)
-                XCTAssertEqual(sut.thrownErrors.count, 1)
+        sut.implementation = .throws(URLError(.badURL))
 
-                let firstThrownError = try XCTUnwrap(sut.thrownErrors.first as? URLError)
+        do {
+            try await invoke(("b", false))
+            XCTFail("Expected invoke to throw an error")
+        } catch let error as URLError {
+            XCTAssertEqual(error.code, .badURL)
+            XCTAssertEqual(sut.thrownErrors.count, 1)
 
-                XCTAssertEqual(firstThrownError.code, .badURL)
-            } catch {
-                XCTFail("Expected \(error) to equal URLError(.badURL)")
-            }
+            let firstThrownError = try XCTUnwrap(sut.thrownErrors.first as? URLError)
 
-            sut.error = URLError(.badServerResponse)
+            XCTAssertEqual(firstThrownError.code, .badURL)
+        } catch {
+            XCTFail("Expected \(error) to equal URLError(.badURL)")
+        }
 
-            do {
-                try await invoke(("c", true))
-                XCTFail("Expected invoke to throw an error")
-            } catch let error as URLError {
-                XCTAssertEqual(error.code, .badServerResponse)
-                XCTAssertEqual(sut.thrownErrors.count, 2)
+        sut.implementation = .throws(URLError(.badServerResponse))
 
-                let firstThrownError = try XCTUnwrap(sut.thrownErrors.first as? URLError)
-                let lastThrownError = try XCTUnwrap(sut.thrownErrors.last as? URLError)
+        do {
+            try await invoke(("c", true))
+            XCTFail("Expected invoke to throw an error")
+        } catch let error as URLError {
+            XCTAssertEqual(error.code, .badServerResponse)
+            XCTAssertEqual(sut.thrownErrors.count, 2)
 
-                XCTAssertEqual(firstThrownError.code, .badURL)
-                XCTAssertEqual(lastThrownError.code, .badServerResponse)
-            } catch {
-                XCTFail("Expected \(error) to equal URLError(.badServerResponse)")
-            }
+            let firstThrownError = try XCTUnwrap(sut.thrownErrors.first as? URLError)
+            let lastThrownError = try XCTUnwrap(sut.thrownErrors.last as? URLError)
+
+            XCTAssertEqual(firstThrownError.code, .badURL)
+            XCTAssertEqual(lastThrownError.code, .badServerResponse)
+        } catch {
+            XCTFail("Expected \(error) to equal URLError(.badServerResponse)")
         }
     }
 
     // MARK: Last Thrown Error Tests
 
     func testLastThrownError() async throws {
-        try await self.withSUT { sut, invoke in
-            XCTAssertNil(sut.lastThrownError)
-
-            try await invoke(("a", true))
-            XCTAssertNil(sut.lastThrownError)
-
-            sut.error = URLError(.badURL)
-
-            do {
-                try await invoke(("b", false))
-                XCTFail("Expected invoke to throw an error")
-            } catch let error as URLError {
-                XCTAssertEqual(error.code, .badURL)
-
-                let lastThrownError = try XCTUnwrap(sut.lastThrownError as? URLError)
-
-                XCTAssertEqual(lastThrownError.code, .badURL)
-            } catch {
-                XCTFail("Expected \(error) to equal URLError(.badURL)")
-            }
-
-            sut.error = URLError(.badServerResponse)
-
-            do {
-                try await invoke(("c", true))
-                XCTFail("Expected invoke to throw an error")
-            } catch let error as URLError {
-                XCTAssertEqual(error.code, .badServerResponse)
-
-                let lastThrownError = try XCTUnwrap(sut.lastThrownError as? URLError)
-
-                XCTAssertEqual(lastThrownError.code, .badServerResponse)
-            } catch {
-                XCTFail("Expected \(error) to equal URLError(.badServerResponse)")
-            }
-        }
-    }
-}
-
-// MARK: - Helpers
-
-extension MockVoidAsyncThrowingMethodWithParametersTests {
-    private func withSUT(
-        test: (
-            _ sut: SUT,
-            _ invoke: (Arguments) async throws -> Void
-        ) async throws -> Void
-    ) async rethrows {
         let (sut, invoke) = SUT.makeMethod()
 
-        try await test(sut, invoke)
+        XCTAssertNil(sut.lastThrownError)
+
+        try await invoke(("a", true))
+        XCTAssertNil(sut.lastThrownError)
+
+        sut.implementation = .throws(URLError(.badURL))
+
+        do {
+            try await invoke(("b", false))
+            XCTFail("Expected invoke to throw an error")
+        } catch let error as URLError {
+            XCTAssertEqual(error.code, .badURL)
+
+            let lastThrownError = try XCTUnwrap(sut.lastThrownError as? URLError)
+
+            XCTAssertEqual(lastThrownError.code, .badURL)
+        } catch {
+            XCTFail("Expected \(error) to equal URLError(.badURL)")
+        }
+
+        sut.implementation = .throws(URLError(.badServerResponse))
+
+        do {
+            try await invoke(("c", true))
+            XCTFail("Expected invoke to throw an error")
+        } catch let error as URLError {
+            XCTAssertEqual(error.code, .badServerResponse)
+
+            let lastThrownError = try XCTUnwrap(sut.lastThrownError as? URLError)
+
+            XCTAssertEqual(lastThrownError.code, .badServerResponse)
+        } catch {
+            XCTFail("Expected \(error) to equal URLError(.badServerResponse)")
+        }
     }
 }
